Rearrange the classes
