+++ GraduationTracker/GraduationTracker.Tests.Unit/GraduationTrackerTests.cs
@@ -2,6 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using System.Linq;
+using GraduationTracker.Models;
 
 namespace GraduationTracker.Tests.Unit
 {
