+++ GraduationTracker/GraduationTracker/GraduationTracker.csproj
@@ -40,8 +40,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Course.cs" />
-    <Compile Include="Diploma.cs" />
+    <Compile Include="Models\Course.cs" />
+    <Compile Include="Models\Diploma.cs" />
     <Compile Include="GraduationTracker.cs" />
     <Compile Include="Repositories\DiplomaRepo.cs" />
     <Compile Include="Repositories\IDiplomaRepo.cs" />
@@ -53,8 +53,8 @@
     <Compile Include="STANDING.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Repositories\Repository.cs" />
-    <Compile Include="Requirement.cs" />
-    <Compile Include="Student.cs" />
+    <Compile Include="Models\Requirement.cs" />
+    <Compile Include="Models\Student.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
