+++ GraduationTracker/GraduationTracker/Models/Course.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GraduationTracker.Models
+{
+    public class Course
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public int Mark { get; set; }
+        public int Credits { get; }
+     }
+}
